added some more logic and more functionally. Was able to create orders with the action for create within the orders controller

+++ MVCPizzaStore/PizzaStore.DataAccess/Pizza.cs
@@ -13,7 +13,7 @@ namespace PizzaStore.DataAccess
         public int Id { get; set; }
         public string Name { get; set; }
         public string CrustType { get; set; }
-        public decimal LinePrice { get; set; }
+        public double LinePrice { get; set; }
         public int? PizzaOrderId { get; set; }
 
         public virtual PizzaOrder PizzaOrder { get; set; }
